feat add Cover component
