+++ app/renderer/src/pages/AlbumList.tsx
@@ -9,8 +9,8 @@ import styled from 'styled-components';
 import { readDB, refreshCover, Track } from '../api';
 import { HorenContext } from '../App';
 import Modal from '../components/Modal';
-import defaultCover from '../defaultCover';
 import Page, { PageProps } from './_page';
+import { AlbumCover } from '../components/Cover';
 
 const ALBUM = styled.ul`
   margin: 0;
@@ -408,15 +408,3 @@ function AlbumItem({ album, onOpen }: AlbumItemProps) {
     </Item>
   );
 }
-
-const AlbumCover = (props: React.ImgHTMLAttributes<HTMLImageElement>) => {
-  const [isError, setIsError] = useState(true);
-  const handleError = (e: React.SyntheticEvent) => {
-    if (isError) {
-      setIsError(false);
-      (e.target as HTMLImageElement).src =
-        'data:image/png;base64,' + defaultCover;
-    }
-  };
-  return <img onError={handleError} {...props} />;
-};
